// src/app/models/movie.interface.ts
export interface MovieSearch {
    Title: string;
    Year: string;
    imdbID: string;
    Type: string;
    Poster: string;
}

export interface Rating {
    Source: string;
    Value: string;
}

export interface MovieById {
    Title: string;
    Year: string;
    Rated: string;
    Released: string;
    Runtime: string;
    Genre: string;
    Director: string;
    Writer: string;
    Actors: string;
    Plot: string;
    Language: string;
    Country: string;
    Awards: string;
    Poster: string;
    Metascore: string;
    imdbRating: string;
    imdbVotes: string;
    imdbID: string;
    Type: string;
    totalSeasons?: string;
    Response: string;
    Ratings: Rating[];
    DVD?: string;
    BoxOffice?: string;
    Production?: string;
    Website?: string;
    listOrder: number;
}